add timeline fixtures and refactor snapshots to new folder structure
