+++ test/e2e/frontend/cypress/fixtures/version-history/snapshots/device-snapshots.json
