+++ test/e2e/frontend/cypress/fixtures/version-history/snapshots/device2-full-snapshot1.json
