+++ test/e2e/frontend/cypress/fixtures/version-history/snapshots/instance-snapshots.json
