+++ test/e2e/frontend/cypress/fixtures/version-history/snapshots/instance2-full-snapshot2.json
