+++ test/e2e/frontend/cypress/fixtures/version-history/snapshots/instance2-snapshot2.json
