+++ test/e2e/frontend/cypress/fixtures/version-history/snapshots/snapshot-with-credentials.json
