+++ test/e2e/frontend/cypress/fixtures/version-history/snapshots/upload-for-download.json
