+++ test/e2e/frontend/cypress/fixtures/version-history/snapshots/upload2.json
