+++ test/e2e/frontend/cypress/fixtures/version-history/snapshots/upload4.json
